refactor: colocate the dicom-synth fixture writers in synthFixtures
