+++ src/applyMappingsWorker.test.ts
@@ -1,7 +1,7 @@
 import { mkdtempSync, readFileSync, rmSync } from 'node:fs'
 import { tmpdir } from 'node:os'
 import { join } from 'node:path'
-import { VALID_CT_IMAGE, writeSynthFile } from '../testutils/minimalDicom'
+import { VALID_CT_IMAGE, writeSynthFile } from '../testutils/synthFixtures'
 import {
   createMappingWorker,
   runMappingWorker,
